Fin V1.4

+++ backend/src/main/kotlin/com/loshermanos/controller/dto/SaleItemDTO.kt
@@ -1,5 +1,5 @@
 package com.loshermanos.controller.dto
 
-class SaleItemDTO(val id : Long,val amount :Long) {
+class SaleItemDTO(val id : Long,val amount :Double) {
 
 }
