+++ backend/src/main/kotlin/com/loshermanos/controller/dto/StockDTO.kt
@@ -1,4 +1,4 @@
 package com.loshermanos.controller.dto
 
-class StockDTO(val id:Long,val amount:Long,val add :Boolean) {
+class StockDTO(val id:Long,val amount:Double,val add :Boolean) {
 }
